USB Gadget HAL implementation
Add implementation for USB Gadget HAL. This HAL is called by
the framework in order to configure the composition of USB
xgadget functions, which had previously been handled by init.rc
property triggers. The HAL implementation provides improved
performance and additional flexibility not offered by Android
init syntax.

Enablement of this HAL is done in the target-specific makefile
and manifest.

Change-Id: I909db39f93e13871d72607cc5de6aec98a383798