Add USB HAL implementation
Add implementation for IUsb (not to be confused with IUsbGadget)
HAL which provides abstraction for querying and notification of
USB port status changes. It is primarily meant for Type-C ports
that support changing roles dynamically.

This implements version 1.1 of IUsb. Unlike the default 1.0
implementation bundled with AOSP, this version supports getting
port status using the Type-C connector class interface found
in upstream kernel.

This code is adopted from
https://android.googlesource.com/device/google/wahoo/+/refs/heads/master/usb/

Change-Id: I43c5e0214989d6ce650fbacb23d1bb0b592f2dfc