USB HAL: Add vendor pickup makefile for self-contained compilation
Normally to enable this component for inclusion in a target's
vendor image, the PRODUCT_PACKAGES of the target's makefile must
include this module. Additionally, for a HAL it must also be
added to the vintf manifest.

To support self-containment of this component, vendor_product.mk
is defined with the above definitions, and is "picked up" by a
target via symlink defined in the top-level manifest. The vintf
xml files can live under /vendor/etc/vintf/manifest as vintf
fragments. Since USB Gadget HAL is optional, this can be
conditionally included.

Change-Id: I91d81ccfe45014777c3e10c93d86d1266fe1e409