USB: Split Gadget HAL into its own service
The HAL service includes both USB HAL and USB Gadget HAL
implementations. Since the latter is optional, split it
out into its own service with its own init_rc and vintf
fragment.

This allows for further simplification of the vendor pickup
makefile as now it can determine whether to include the
Gadget HAL by conditionally including it in PRODUCT_PACKAGES
on a per-target basis. With the vintf_fragments specified in
Android.bp, the use of PRODUCT_COPY_FILES can now be removed.

Change-Id: I5cd060ba704690d8c98de36322283b139c99f134