USB: Configure USB Diag over FFS
On targets where diag-router is used, override the
vendor.usb.diag.func.name property to 'ffs' to
instantiate the Diag function instances to use
F_FS instead of the legacy f_diag driver. This
is done using PRODUCT_PROPERTY_OVERRIDES in which
this property will be placed in /vendor/build.prop.

Change-Id: I529a081a0d6988628944a9020b61c061baa877a4