USB: Configure USB MTP over ffs for kernels above 5.4
On targets using 5.4 kernel and above the MTP is supported
over ffs and the legacy f_mtp is deprecated. So, set
the vendor.usb.use_ffs_mtp to 1 so that ffs based mtp
is used. This is done using PRODUCT_PROPERTY_OVERRIDES
in which the properties will be placed in /vendor/build.prop.

Change-Id: I2080aab9b5c79fe426db4dc18dc80b5b78b80317