USB: HAL: Prevent overriding the MaxPower value to zero
If multiple uevents are generated when the device is in
power delivery mode, the saved state of MaxPower is overwritten
to zero. This causes the MaxPower to be set to 0 when a device
that does not support power delivery is connected.
Fix this by ignoring the uevents if the power operation mode
is ot changed.

Change-Id: I1414e2328e4cada640be6eb0d89a1cafa6eaf92f