USB: HAL: Allow device suspend when boots up with hub connected
The current implementation to support device suspend with hub
connected does not covers the case where the device boots up
with hub connected. Fix this by checking all the devices connected
as part of setCallback and depending on the type of interfaces
supported by the device, enable the autosuspend as well as wakeup
capability.

Change-Id: I1fbd1090a68e239e49dfbfac16fe529e29bd4605