init.qcom.usb.rc: Prevent enumeration with incorrect controller
On the latest kernels, the DUMMY_HCD driver is enabled and is
causing the init to populate the sys.usb.controller with
dummy_hcd.0 and which is overwritten by the init.target.rc.
But this race has caused the failure of enumeration of persist
compositions after boot-up.

Fix this by enusring that the sys.usb.configfs is set only when
a proper vendor.usb.controller is set, so the required rc rules
are run only after the proper vendor.usb.controller and
sys.usb.controller is populated.

Change-Id: I5767fe5ab16276fe392e82d83682a5cf38f9d9bb