USBGadget HAL: Expand support for overriding default composition
When ADB-only composition is chosen, persist.vendor.usb.config
is checked to allow for overriding with a custom composition.
Instead of calling SetProperty to handle it from initrc, handle
this directly in the HAL by tokenizing the string and looking
up each function to link to the composition. Finally, look up
the combination to find the supported VID/PID.

Currently the supported combinations is maintained as a static
lookup table (std::map), but in the future if this needs to
grow it can be enhanced to be kept in a data file (similar to
fstab) in the filesystem.

Change-Id: Id71f56d02b2fcf1663598eba47d2f7cd52cea429