init.qcom.usb.rc: Prevent rc rules if USB Gadget HAL is enabled
If Gadget HAL is enabled prevent rc rules from running for sys.usb.config
triggers by preventing the sys.usb.configfs to be set to 1 and setting it
to 2.
This is done by using the new property vendor.usb.uses_gadget_hal which
signifies whether Gadget HAL is enabled or not for the platform. Also,
enable compostition switch by starting usbd when user sets the
vendor.usb.config if Gadget HAL is enabled.

Change-Id: I946813109cdb4613fb6695fad58c84388fcde50b