USB: Add default property overrides for common functions
Several of the vendor.usb.* properties for function settings
such as RmNet, RNDIS, DPL and QDSS are set in many targets'
init.target.rc files and often share the same defaults. Reduce
the duplication by definining common defaults here in this
component as PRODUCT_PROPERTY_OVERRIDES so that they get generated
in /vendor/build.prop. This still allows them to be further
overridden in respective init*.rc files if needed.

Change-Id: If2732cf1410d12e9032ecdcafa7cc8d12ed81083