init.qcom.usb.rc: Remove os_desc symlink for compositions w/o os descs
Configurations which include functions that don't implement OS
descriptor support will fail to enumerate with a Windows host.  If an
os_desc symlink is present, the device will respond back to the PC w/
an invalid OS desc response, leading to enumeration failures.  Remove
the symlink for these compositions, and re-add when a composition
change occurs.

Change-Id: I06c23f72b4a3cf023666e84fc64601fffa1689f7