USB: Configure Diag MDM instance on 90D9
On certain platforms like IVI where diag-router is used, we do
not expect an MDM to be attached in case of 90D9. Now to be able
to collect WLAN FW diag logs, we need to use the diag_mdm2
instance. So use "vendor.usb.diag_mdm.inst.name" property to
decide which instance name to bind, while keeping the default
value to "diag_mdm".

Change-Id: I53ddb2ab3540b3202c50927bf9c0e6b3e0434b0b