USBGadgetHAL: Support QDSS debug interface
Some compositions additionally specify the "enable_debug_inface"
option for the QDSS function. This results in the driver exposing
a control interface in addition to the debug trace interface.
For the handful of PIDs that enable this, distinguish them by
substituting "qdss_debug" instead of "qdss" and taking advantage
of the new "actual_order" entry in the supported_compositions map.

Change-Id: I5058f2554f5e1556f8f2627604a95eebca3c5bea