init.qti.usb.debug: Only enable ftrace on non-perf kernels
Only enable USB ftrace events on non-perf images in order to
reduce memory overhead of the additional tracing instances.
Borrow the same check from kernel scripts which check for the
presence of the msm_rtb module.

Also remove the additional "&& property:ro.build.type=userdebug"
condition on the init trigger since this is already implied
with the vendor_product.mk pickup rules which determine the
inclusion of the debug rc/sh in the build, and additionally
allow it to work on "eng" build variants.

Change-Id: I3071717d20fafb94e75bc29af68e41e87d90f17a