Add minimal USB init RC for QMAA
To facilitate early bringup on pre-silicon platforms, introduce
init.qti.usb.qmaa.rc which is a stripped-down init script that
performs only the necessary configuration to bring up ADB over USB.
This avoids unnecessary dependencies on Android init and other
framework services which may not be stable.  Note the ADB-only
composition is supported at this time.

This is generally enabled when the global TARGET_USES_QMAA flag
is set to true.  It can be overridden (that is, enable the full
USB scripts, HALs, etc) when TARGET_USES_QMAA_OVERRIDE_USB is also
set to true.

Change-Id: Idd39efafdbaaa8a9f407abc04b4270fdfa15960e